+++ src/Controller/CalendarController.php
@@ -86,7 +86,7 @@ class CalendarController extends CommonController
 
         $tpl_params['months'] = array_fill(1, 12, ['nb_images' => 0]);
 
-        $intl_date_formatter = new \IntlDateFormatter($appUserService->getUser()->getUserInfos()->getLanguage(), \IntlDateFormatter::FULL, \IntlDateFormatter::NONE, null, null, 'MMM');
+        $intl_date_formatter = new \IntlDateFormatter($request->get('_locale'), \IntlDateFormatter::FULL, \IntlDateFormatter::NONE, null, null, 'MMM');
         for ($month = 1;$month <= 12;$month++) {
             $tpl_params['months'][$month]['label'] = $intl_date_formatter->format(new \DateTime(sprintf('2022-%d-01', $month)));
         }
@@ -132,7 +132,7 @@ class CalendarController extends CommonController
         $tpl_params['month_date'] = new \DateTime("${year}-${month}-01");
         $tpl_params['number_of_images'] = 0;
 
-        $intl_date_formatter = new \IntlDateFormatter($appUserService->getUser()->getUserInfos()->getLanguage(), \IntlDateFormatter::FULL, \IntlDateFormatter::NONE, null, null, 'MMMM');
+        $intl_date_formatter = new \IntlDateFormatter($request->get('_locale'), \IntlDateFormatter::FULL, \IntlDateFormatter::NONE, null, null, 'MMMM');
         $tpl_params['month_label'] = $intl_date_formatter->format($tpl_params['month_date']);
 
         $tpl_params['days'] = array_fill(1, (int) $tpl_params['month_date']->format('t'), ['nb_images' => 0]);
@@ -184,7 +184,7 @@ class CalendarController extends CommonController
         $tpl_params['day'] = $this->formatDatePart($day);
         $tpl_params['current_day'] = new \DateTime("${year}-${month}-${day}");
 
-        $intl_date_formatter = new \IntlDateFormatter($appUserService->getUser()->getUserInfos()->getLanguage(), \IntlDateFormatter::FULL, \IntlDateFormatter::NONE, null, null, 'MMMM');
+        $intl_date_formatter = new \IntlDateFormatter($request->get('_locale'), \IntlDateFormatter::FULL, \IntlDateFormatter::NONE, null, null, 'MMMM');
         $tpl_params['month_label'] = $intl_date_formatter->format($tpl_params['current_day']);
 
         $thumbnails = [];