+++ src/Controller/MenubarController.php
@@ -14,16 +14,20 @@ namespace App\Controller;
 use Phyxo\Conf;
 use Phyxo\MenuBar;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\RequestStack;
 use Symfony\Component\HttpFoundation\Response;
 
 class MenubarController extends CommonController
 {
-    public function navigation(Request $request, Conf $conf, MenuBar $menuBar): Response
+    public function navigation(RequestStack $requestStack, Conf $conf, MenuBar $menuBar): Response
     {
         $tpl_params = [];
 
-        $tpl_params = array_merge($tpl_params, $menuBar->getBlocks());
-        $tpl_params = array_merge($tpl_params, $this->loadThemeConf($request->getSession()->get('_theme'), $conf));
+        $tpl_params['blocks'] = $menuBar->getBlocks();
+        $tpl_params['current_route'] = $requestStack->getMainRequest()->get('_route');
+        $tpl_params['current_route_params'] = $requestStack->getMainRequest()->get('_route_params');
+
+        $tpl_params = array_merge($tpl_params, $this->loadThemeConf($requestStack->getSession()->get('_theme'), $conf));
 
         return $this->render('_menubar.html.twig', $tpl_params);
     }